third_party/WebKit/LayoutTests/platform/mac-mac10.10/virtual/mojo-loading/http/tests/inspector/bindings/livelocation-main-frame-navigated-expected.txt

Verify that UISourceCodes are removed as main frame gets navigated.


Running: attachFrame
[ CREATE ]  LiveLocation-js: http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-typescript.ts:0:0
[ CREATE ]  LiveLocation-css: http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-sass.scss:0:0

Running: navigateMainFrame
[ UPDATE ]  LiveLocation-css: null
[ UPDATE ]  LiveLocation-css: null
[ UPDATE ]  LiveLocation-js: http://127.0.0.1:8000/inspector/bindings/resources/sourcemap-script.js:0:0
[ UPDATE ]  LiveLocation-js: debugger:///VM53 sourcemap-script.js:0:0
Page reloaded.
[  GET   ]  LiveLocation-js: null
[  GET   ]  LiveLocation-css: null

Running: navigateBack
Page reloaded.
